llm_tool_youtube_video_agent/requirements.txt
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
groq